requirements.txt
bcrypt==3.2.0
fastapi==0.75.0
passlib==1.7.4
PyJWT==2.3.0
python-multipart==0.0.5
uvicorn==0.17.6